Add `language_models` crate to house language model providers
