+++ crates/language_model/src/language_model.rs
@@ -1,23 +1,19 @@
-pub mod logging;
 mod model;
-pub mod provider;
 mod rate_limiter;
 mod registry;
 mod request;
 mod role;
-pub mod settings;
+
+#[cfg(any(test, feature = "test-support"))]
+pub mod fake_provider;
 
 use anyhow::Result;
-use client::{Client, UserStore};
 use futures::FutureExt;
 use futures::{future::BoxFuture, stream::BoxStream, StreamExt, TryStreamExt as _};
-use gpui::{
-    AnyElement, AnyView, AppContext, AsyncAppContext, Model, SharedString, Task, WindowContext,
-};
+use gpui::{AnyElement, AnyView, AppContext, AsyncAppContext, SharedString, Task, WindowContext};
 pub use model::*;
-use project::Fs;
 use proto::Plan;
-pub(crate) use rate_limiter::*;
+pub use rate_limiter::*;
 pub use registry::*;
 pub use request::*;
 pub use role::*;
@@ -27,14 +23,10 @@ use std::fmt;
 use std::{future::Future, sync::Arc};
 use ui::IconName;
 
-pub fn init(
-    user_store: Model<UserStore>,
-    client: Arc<Client>,
-    fs: Arc<dyn Fs>,
-    cx: &mut AppContext,
-) {
-    settings::init(fs, cx);
-    registry::init(user_store, client, cx);
+pub const ZED_CLOUD_PROVIDER_ID: &str = "zed.dev";
+
+pub fn init(cx: &mut AppContext) {
+    registry::init(cx);
 }
 
 /// The availability of a [`LanguageModel`].
@@ -184,7 +176,7 @@ pub trait LanguageModel: Send + Sync {
     }
 
     #[cfg(any(test, feature = "test-support"))]
-    fn as_fake(&self) -> &provider::fake::FakeLanguageModel {
+    fn as_fake(&self) -> &fake_provider::FakeLanguageModel {
         unimplemented!()
     }
 }
