+++ crates/language_models/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "language_model"
+name = "language_models"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,37 +9,41 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/language_model.rs"
-doctest = false
-
-[features]
-test-support = [
-    # "language/test-support",
-    # "text/test-support",
-]
+path = "src/language_models.rs"
 
 [dependencies]
 anthropic = { workspace = true, features = ["schemars"] }
 anyhow.workspace = true
-base64.workspace = true
+client.workspace = true
 collections.workspace = true
+copilot = { workspace = true, features = ["schemars"] }
+editor.workspace = true
+feature_flags.workspace = true
+fs.workspace = true
 futures.workspace = true
 google_ai = { workspace = true, features = ["schemars"] }
 gpui.workspace = true
 http_client.workspace = true
-image.workspace = true
-log.workspace = true
+language_model.workspace = true
+menu.workspace = true
 ollama = { workspace = true, features = ["schemars"] }
 open_ai = { workspace = true, features = ["schemars"] }
-parking_lot.workspace = true
+project.workspace = true
 proto.workspace = true
 schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+settings.workspace = true
 smol.workspace = true
 strum.workspace = true
+telemetry_events.workspace = true
+theme.workspace = true
+thiserror.workspace = true
+tiktoken-rs.workspace = true
 ui.workspace = true
 util.workspace = true
 
 [dev-dependencies]
-gpui = { workspace = true, features = ["test-support"] }
+editor = { workspace = true, features = ["test-support"] }
+language_model = { workspace = true, features = ["test-support"] }
+project = { workspace = true, features = ["test-support"] }
