+++ crates/language_models/src/provider.rs
@@ -1,8 +1,6 @@
 pub mod anthropic;
 pub mod cloud;
 pub mod copilot_chat;
-#[cfg(any(test, feature = "test-support"))]
-pub mod fake;
 pub mod google;
 pub mod ollama;
 pub mod open_ai;
